Extract ProjectConfigOptions from cli/index.ts
